moved makers into package, implemented grandma, updated abstract CookieMaker
