+++ src/main/java/makers/Grandmas.java
@@ -1,0 +1,12 @@
+package makers;
+
+public class Grandmas extends CookieMaker {
+    public Grandmas () {
+        super();
+        cookieRate = 1;
+        priceForNext = 100;
+        name = "grandma";
+        updateCPS();
+    }
+
+}
